requirements.txt
fastapi
uvicorn
pillow
httpx
opencv-python-headless
numpy
boto3
python-dotenv